reorganize files to minimize use of __init__.py files
